style: New dynamic design for settings screen
changes:
- New design compatible with mobile and desktop/web
- Full code refactor to improve readability/maintainability by separating
  each settings section into a separate file (e.g., appearance section as
  appearance_settings.dart).
- Remove theme_tiles_picker.dart and paste its content into the related
  file (appearance_settings.dart)
- Remove SafeArea from parent; now managed from showModalBottomSheet()
  function.
- Rename the widget and file from settings_sheet to settings_screen.
