+++ lib/app/settings_controller.dart
@@ -5,6 +5,7 @@ import 'package:flutter/foundation.dart' show ChangeNotifier, debugPrint;
 import 'package:flutter/material.dart' show ColorScheme, ThemeMode;
 import 'package:my_quran/app/models.dart';
 import 'package:my_quran/app/services/settings_service.dart';
+import 'package:my_quran/app/widgets/settings/settings_screen.dart';
 import 'package:wakelock_plus/wakelock_plus.dart';
 
 class SettingsController extends ChangeNotifier {
@@ -12,6 +13,7 @@ class SettingsController extends ChangeNotifier {
 
   final SettingsService settingsService;
 
+  int _section = SettingSection.appearance.index;
   String _language = 'ar';
   FontFamily _fontFamily = FontFamily.rustam;
   FontWeight _fontWeight = FontWeight.w500;
@@ -28,6 +30,7 @@ class SettingsController extends ChangeNotifier {
 
   // ── Getters ──
 
+  int get section => _section;
   bool get supportsDynamicColor => _supportsDynamicColor;
   AppTheme get appTheme => _appTheme;
   bool get keepScreenOn => _keepScreenOn;
@@ -48,6 +51,11 @@ class SettingsController extends ChangeNotifier {
       fontFamily == FontFamily.rustam ? FontWeight.w500 : _fontWeight;
 
   // ── Setters ──
+  
+  set section(int value) {
+    _section = value;
+    notifyListeners();
+  }
 
   set appTheme(AppTheme value) {
     _appTheme = value;
